+++ web/apps/labelstudio/src/pages/Settings/StorageSettings/StorageFormNew/Steps/index.ts
@@ -1,0 +1,5 @@
+export { Stepper } from "./Stepper";
+export { ProviderSelectionStep } from "./ProviderSelectionStep";
+export { ProviderDetailsStep } from "./ProviderDetailsStep";
+export { PreviewStep } from "./PreviewStep";
+export { ReviewStep } from "./ReviewStep";
