+++ web/libs/ui/package.json
@@ -4,5 +4,12 @@
   "license": "MIT",
   "private": true,
   "main": "src/index.ts",
-  "files": ["src/tailwind.css", "src/shad/components", "src/ui/assets"]
+  "files": [
+    "src/tailwind.css",
+    "src/shad/components",
+    "src/ui/assets"
+  ],
+  "dependencies": {
+    "@radix-ui/react-radio-group": "^1.3.7"
+  }
 }
