+++ web/libs/ui/src/index.ts
@@ -31,3 +31,5 @@ export * from "./utils/utils";
 // TODO: Remove when DIA-2142 and DIA-2175 are delivered
 export * from "./shadcn";
 export * from "./utils/utils";
+
+export { RadioGroup, RadioGroupItem } from "./shad/components/ui/radio-group";
