+++ web/libs/ui/src/shadcn.ts
@@ -1,4 +1,5 @@
 /// Raw shadcn components for re-export
-export { Badge, type BadgeProps, badgeVariants } from "./shad/components/ui/badge";
-export { Skeleton } from "./shad/components/ui/skeleton";
-export { cn } from "./utils/utils";
+export * from "./shad/components/ui/badge";
+export * from "./shad/components/ui/skeleton";
+export * from "./shad/components/ui/radio-group";
+export * from "./utils/utils";
